requirements.txt
python-telegram-bot==21.6
playwright>=1.50.0
greenlet>=3.2.0
python-dotenv==1.0.1
